ml-service/requirements.txt
# Web Framework
Flask==2.3.2
Flask-CORS==4.0.0
gunicorn==21.2.0
Werkzeug==2.3.6

# Database
psycopg2-binary==2.9.6
SQLAlchemy==2.0.19

# Machine Learning
scikit-learn==1.2.2
numpy==1.24.3
pandas==2.0.3
joblib==1.2.0

# Async Tasks
celery==5.3.1
redis==4.5.5

# HTTP Requests
requests==2.31.0

# Configuration
python-dotenv==1.0.0

# Monitoring & Metrics
prometheus-client==0.17.1

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0

# Code Quality
black==23.7.0
flake8==6.0.0
isort==5.12.0
